UI_WEB: Slider bars are at a working state
Just need to have signals put in when the time comes to save the new
distribution values, etc. Thinking about hiding the sliders unless they're 
explicitly asked for, as there's no way to explicitly declare what
distribution you want.

There's two problems thus far, however:

1) As the steps are set to be 1, if you are actually sliding the slider bar
instead of just clicking within the range, and one of the other elements
is at 0, that element will most likely stay at 0. This has to do with the 
rounding algorithm, we'll get two fractions and neither will have a ceiling
higher than 0 when doing the newValues portion, which means the vector
will come out to (0, 1) and only one of the other elements will break free
from 0. (Tested with three elements). Possible solutions? Set a min of 1,
with the ability to check off vendors that you don't want. Seems easy. 
Can't really think of another way.

2) The distributions will never end up the same at the same point on the 
range for a given slider. If you slide the top bar to 100, making the other
elements 0, when you slide that same bar to 50, it's a relative toss-up as to
what the other bars will be (seems to distribute within a tolerance of about
5%, however). I suspect that particular case could be remedied with the 
inclusion of the 0-disabled solution, but this remains true for the other
points on the line as well. I think it's okay, though; we can't moderate the 
movements of the slider bar (maybe increasing the step to 2 may help,
though), so it may not be something to worry about.