UI_WEB: Working on adding eth vendors
So far the page adds the vendors correctly, as well as propely decrements
the maximum values of the ranges, etc. etc. Need to get delete working, 
then back end stuff, then get the sliders to work.

There's a curious problem right now with the hand-modified input boxes:
after they've been modified once, they won't ever again reflect any changes
to their value except through manual modification. However, the values
will be updated, as well as the innerHTML, so I honestly have no idea why.
It may be that the function that gets called to modify the elements 
dynamically is ending before .value or setAttribute are done modifying the
element, but you can inspect the DOM after moving a slider and the values
seem to be correct. Will investigate later, trying to get the functionality set 
up so that all I have to do is fix UI quirks and then wash my hands of this
madness.