UI_WEB: Little bit of debugging to fix editProfile
Made a change to editHoneydProfile that propagated to another page
that uses the same jade file but in a different way. Fixed it so that it 
works now, but eventually we need to split the add profile and edit 
profile pages up into two and actually get editing of profiles working.