UI_CORE: Added a function, fixed a loop
Added the NumberOfChildren function for NodeManager to prevent
non-leaf nodes from having nodes created. Before, every run would
produce a node of the "Linux Linux 2.6.X general purpose" profile,
which is just used as a parent node for the profiles that actually contain
personality values, etc. Now, only leaf nodes will be created.

Also, the segfault problem where the NodeProfile pointer "p" in 
GenerateNodes was becoming a null pointer and then subsequently
being accessed was due to a faulty loop iterator assignment. In the 
MAC Counter loop, if all of the vendors had m_counts less than 0, 
the loop would kick out without assigning curNode a proper NodeProfile.
To fix this, the loop will only terminate when the else condition inside of it
is met properly, resulting in an assignment to curNode.