UI_CORE: HoneydConfiguration auto-finds interfaces on instantiation
When instantiating the HoneydConfiguration object, it now loads the 
subnets table with both the XML interfaces and found physical interfaces.
If there is an intersection between the XML and the found interface, the
XML value is used; if a subnet is found that is not in the XML, it is added
to the subnet table.