UI_WEB: Testing Profile editing
Because of the way that the code-behind for the WEB UI works, editing
profiles is going to be a new kind of monster. UpdateProfile(string) relies
on the edited information already being inside the profile and the 
hash map, whereas the WEB UI doesn't put the information into the 
profile structs within the table, it just makes a copy and adds it (that's why
adding profiles works). In order to edit, the web UI is going to have to 
replace the existing profile with the new profile generated from the client
side.