UI_WEB: Fixing EditProfile name change bug
Previously, when one attempted to edit profiles through the web UI, 
if the name was changed the previous profile would be left as it was and
a new profile would be created with the new name.

This calls RenameProfile instead of checking the return value of AddProfile and 
then calling UpdateProfile.