UI_CORE && UI_WEB: Fix some profile editing bugs
UI Core had some problems with not updating the defaul profile's ptree, 
which would lead to multiple ptree children being created when 
UpdateProfileTree was called (it never modified the default ptree, leaving
all of the previous children in the xml and adding another set onto it). Gave
default it's own case, seems to work, the correct profile was changed and 
there's only one ptree node for each profile. 

Web was doing the editing wrong, I was changing the profile and then 
calling Rename, which would fail because I was trying to rename a profile
to itself. Fixed that, works with seemingly no problems.