UI_WEB && UI_CORE: Old bug fix made new bug in profile editing
Had to add an argument for the SaveProfile function to help determine
whether to follow AddProfile workflow or EditProfile workflow. 

UI_Core function had a faulty conditional for the "default" case in 
UpdateProfileTree, s.t. it was still never getting hit and thus duped
profiles occured in profiles.xml and lingering profiles that should've been
deleted showing up in the WEB UI as a result.