UI_Core: NodeManager - fix deadlock in node adjustment function
Previously we assumed there would be at least as many nodes as counters,
this simply isn't the case and when that occurs there would be a never
ending chain of under and over populated counters to shift to and from.

To fix this we now do the following
1. Only profiles with at least 1 node can be considering overpopulated
2. Only profiles who deserve at least 1 node based on the distributions and
number of nodes can be considering underpopulated.
3. When checking mac vendors, we skip profiles without nodes.
4. When checking port allocation, we skip profiles without nodes.

Testing on the xml that produced the original case seems to show that the
problem has been resolved by these changes.