HHCONFIG: Uncaught ptree error now caught
From what the header says, there's bad_path and bad_data in terms
of exceptions. Both are now accounted for. I'm going to go through the 
xml_parser exceptions as well and see what needs to be in there.

Put a generic exception catch in there for now, if you see the Unknown 
Exception phrase occur, make another ticket. But it should die gracefully
now.

Issue #305