UI_WEB: Getting hhconfig output onto console
Worked yesterday, but the method in main.js was lost, and after I remade
it (I'm pretty sure) almost exactly as it was it doesn't seem to spawn the 
process at all. The spawn call is made, and then the onExit signal gets hit
with the process returning exit code 0; however, none of the actual 
processing that the autoconfig does actually gets done. I run the 
autoconfig on its own with the same args, and it works.