UI_WEB: HHCONFIG writes to console.log now
Got it to a point where the stdout from nmap is subsumed by the stdout
of the autoconfig, which is then spat out by the client side console. 

Going to work on getting it to either parse nmap output with the 
--stats-every flag enabled (provided ReadableStream class allows
something like that) or (at the very least) just print the stdout of 
autoconfig to the page. Going to look at ReadableStream tomorrow.