UI_WEB: Convert addPort in commonHoneydProfile.jade from innerHTML to DOM object manipulation
Before, the addPort method in the commonHoneydProfile.jade file
was constructing the port rows by just using strings and then appending 
the row string to the innerHTML of the portTable. This had the unfortunate
side-effect of redrawing the entire portTable DOM object, which would
result in changes to ports being lost. Now, all of the row construction has
been converted to using document.* methods, which should allow
for adding ports without losing changes.

Issue: #294