LIB: HoneydConfiguration - error handling and possible fix
Having nested function calls can have undefined behavior depending on the
order of execution when using these hashmaps, using a temporary string
when doing a lookup can cause that string to be unitialized and an empty
or incorrect value used for the lookup. Separated the function calls and
added error handling in case the port really doesn't exist, which is another
problem.