HHCONFIG: Fixing empty ports XML
There was a problem in the RecursiveAddAllPorts method where the loop
to go through all of a profile's ports and check for available scripts was
being terminated early, leading to empty port entries in the XML, which
would then force a quit on a call to LoadAllTemplates. Not any more.