UI_WEB: Problem saving profile information
The addOrEdit boolean I created to help distinguish between a request
to add or edit a profile for the Save function in HoneydProfileBinding
was borked. It was being passed to the c++ as a string instead of a bool,
thus whenever the cvv8 parsing would be done it would get a 1.

So, editing a profile meant adding a profile, and since the name would 
already be present in the hash map it wouldn't do anything. 

Fixed that, going to go through and put some error checking around 
calls to HoneydConfiguration within the web UI c++ as well as inside 
HoneydConfiguration itself.