UI_WEB: Merge conflict from long ago removed needed method
clearCreated moderated the shown vendors when inheritanceChanged is 
clicked. That was lost somewhere, it's back and it seems to work fine.

In addition, port deletion wasn't working, the button type wasn't being
set properly in the DOM version of port row creation. It is now, works fine.

"default" profile can no longer be renamed.

Issue: #332