UI_WEB: Empty key when going to add profile
** Should solve a similar problem for the edit page **

If you went to addHoneydProfile without a query string 
(i.e. https://localhost:8080/addHoneydProfile instead of the needed
https://localhost:8080/addHoneydProfile?parent=blah)
it would crash from an emptyKeyException. There was no keyExists check
in the GetPorts method, so it would try and access m_profiles with 
a null string, leading to the crash. Now, it'll check for key existence, 
and if there's failures down the line it will pop an alert box and redirect
to the configHoneydProfiles page.

Issue: #334