LIB: AddNewNodes not allowing bulk node creation
The AddNewNodes function would allow creation of nodes of the same 
profile, but only if one was created at a time; this was a result of the calls
to AddNewNode within AddNewNodes getting passed the same MAC 
address, which would then create the name of the node that was just
created, which would disallow creation of more than one node. Should
be fine now.