UI_WEB: Getting the autoconfig into the first run wizard
A rudimentary version of this is in place, just adds another button and 
allows the user to select which path to follow, as it were. Might gussy it up
later but in terms of exposing the functionality all that really needs to be
done is the stuff in issue #327.

Issue: #277