LIB && UI_WEB: Added profile dist support for web UI, and updateNodeMAC
Before, when a profile's ethernet vendors were changed, the nodes that
were associated with it didn't have their MAC addresses updated to reflect
the new vendors. Now, a rudimentary version is in place. Takes advantage
of the vendor distributions differently than NodeManager but it should
produce similar results after fine-tuning.

Also added some distribution carryover to stop an edited profile from
being written to XML with a 0 distribution. Added profiles will be written 
that way, but that's fine until a better solution is found.

Issue: #267