UI_WEB && LIB: Adding 'generated' profile tag support, fix bug
Generated wasn't being used in the web UI before, it is now. 

There was a bug for the generated profiles where if any were edited, 
there would be duplicates of each inside of the xmls. This was a result of 
the UpdateProfileTree method not clearing the <profiles> section of the
ptree when updating. Fixed.