MULT: Enable No AUTH or mid-AUTH connections for curl SMTP
Before, the curl library  calls were hardcoded to only accept and use
SSL/TLS connections. Now, it's a user configurable option. If
the user wants to send email alerts through a server that only allows
AUTH PLAIN/AUTH LOGIN they merely have to supply the same 
credentials as before, plus make sure that the Use SMTP Authentication
box is checked in the Web UI or that the value is 1 in the NOVAConfig.txt
file.

Enabling SMTP Auth means that curl will first try to establish the SSL/TLS
connection, and if that fails, attempt to use the username password
combination for AUTH LOGIN. If this, too, fails, there's no recourse, the
message is not sent.