QUASAR: Adding test code for Mothership
Juuust some testing code for using socket.io for bidirectional RPC. 
Wrote a quick, lame little client that can start and stop Novad and the 
Haystack. Going to extend that out a bit, get a feel for the difficulty of 
getting suspects and having two way connections, and then this thing
should start chuggin along.

Issue #356 (if only it could've been 256 *sad face* )