QUASAR: Add some 'do you want to restart haystack?' prompts
Still needs some more; this is just the easy ones where we used nowjs calls.
It's harder with form POSTs since the changes haven't taken affect when
you're sumbitting the form. We could probably just change these to nowjs
calls, or have a hidden form value for if you want to restart honeyd or not.

Ticket #364