QUASAR: Transitioned to use of websockets for Mothership messaging
Switched from socket.io to websockets because it offers an easier and 
more controlled way of implementing the structure that we need. 

Thus far the Quasar communication server (or whatever you want to call
it) can accept 5 message types: Start/Stop Novad, Start/Stop Haystack 
and WriteSetting, which takes args. It responds to all of these messages in
an unvalidated, rudimentary way (i.e. it doesn't check for success, it just 
responds with a success message after the requisite functions have been 
processed. 

I'm going to try to emulate a hostile suspect being received, just to get a 
good idea for what kind of JSON structure will be required to pass it, as 
well as the size for said object, and then start thinking about getting 
wss:// connections instead of ws://

Issue #356