QUASAR: Add wss:// for mothership communication, reconnect ability
Quasar will now try to reconnect with the mothership every 30 seconds 
until it succeeds; when it successfully connects, it will stop trying until
the next connection problem.

Sends out a message on the mothership websocket now if there's a 
hostile suspect found.

server-to-server communication now uses wss:// instead of ws://, or, in 
other words, secure websockets instead of regular websockets. Uses the
same key and certificate file as express for now.

Issue: #356