MOTHERSHIP: Prevent error from being thrown if page containing UpdateConnectionsList isn't loaded
Before, if a client connected and there wasn't a jade file up with a defintion
for UpdateConnectionsList, an error would be thrown which would break 
the connection and essentially repeat itself ad nauseum until the client 
connected in favorable conditions. Now, it's a little more general, 
the Mothership side won't attempt to call UpdateConnectionsList if there
isn't a definition for it.