MOTHERSHIP: Autoconfig client selection elements
When a client is selected, a div is created that contains a label of the 
client's name, an input of the number type for the number of nodes to 
create, and a drop down select box of the interfaces for the given client
(just a hardcoded dummy value for testing, currently). This is subject to 
change with the addition of the node creation ratio, as opposed to a static
number of nodes per client, but as that feature isn't yet complete, the page
is made to the current specifications for the haystack autoconfig tool.