MOTHERSHIP: Page-agnostic suspect updates to mothership
Now, the mothership will receive suspect updates regardless of whether
Quasar is on the main.jade page. Torn between whether to maintain a 
buffer of hostile events that occurred while the mothership wasn't on a
suspectGrid page, and merely immediately render them on grid load,
or just query every client for hostile suspects on grid load. The latter was
easier, so that's how it's done now, but I'd like to change it later.