MOTHERSHIP: Missing index suborning intent, fixed now
Now, the Hostile events button should show an aggregate count of 
off-main-page events, and the connection monitor should show a discrete
count of events seen for each suspect. Note that the relationship between
these counts is necessarily minimal; with the way that the same IP can come
from multiple clients, and clients going from hostile to benign all the time,
the aggregate count is not made to be the sum of the client event counts
or something like that.