MOTHERSHIP: set framework for viewing suspect details in mothership
Not done, solely because the best way to do it isn't yet clear. Could do 
something as simple as query the client to get the requested information,
but then you have to block on the page render until a response was 
received with the desired information; could also get details string
on receipt of hostile suspect events, but that's a big string to be 
constantly sending over the network, though it allows the details page to be
rendered at will.