MOTHERSHIP: Cleaning and repopulating main.jade on browser reconnect
If you close down the server and bring it up back up within a small amount
of time, the browser object will reconnect to the server. Since mothership
will be run with forever, and thus be subject to this state in a failure,
I added some support on main.jade for browser reconnects (i.e. cleaning
all dynamically populated elements on closure/restart, and repopulating
them after receipt of the startup messages)