MOTHERSHIP: Standardize use of single quotes in mothership, shift loading non-essential javascript
All outside quotes are single, inside are double. In addition, a majority of 
the javascript not essential to the page loading were moved to 
individual js files that will load with DOM, not block the DOM before they're
loaded. I see a noticeable increase in page load times, especially on the 
main.jade GET, where it could take upwards of a second to load completely