MISC: nova_rsyslog_helper script for remote logging
I don't know how useful this is going to be in the long run, considering we use
the syslog log levels and anyone who has rsyslog set up will most likely 
already be grabbing all the logs at the levels they need anyways. But this will
specifically target a remote rsyslog instance and place it within the 
40-nova.conf file. There will need to be requisite UI work done, as well as 
changes made to a user's rsyslog server; because we use the expression
filtering, and because remote logging does not allow you to be more 
specific than the destination for the log, rules will have to be put in place on
the rsyslog server's side that routes the Nova-specific logs to whatever
place they're wanted.