INSTALLER: Test new rsyslog enabling structure
Just made the script put a new file inside /etc/rsyslog.d/, instead of editing
on that's in there. Still needs sudo permissions but it's a little more 
indirect in its interaction now. The last format for the filter wouldn't work
when it was in the same file as the other, local logging stuff was, so two
birds, one stone.