MISC: Rsyslog should be good now
Cps the file into the rsyslog.d directory and then edits it. Trying to grab it 
from the .config directory or what have you was a nightmare due to $USER
variable, so it's just going to go into /usr/share/nova/ for now. It doesn't 
get edited inside that folder, just copied out of it into /etc/rsyslog.d/