QUASAR: Shift click completed for configHoneydNodes
With this commit, I believe we can safely say that issue #349 in its original
form is closed. There is some more work that I'm going to do (such as 
adding a "Delete All" button) but every node that will be selected within the
viewable grid is able to be deleted as they were before. 

Issue #349