QUASAR: Delete By Profiles button and fixing problem on main.jade grid
The grid on main.jade keys on a string that represents the href to the 
suspect details page; this conflicted with the way the configHoneyd grid
is set up (admittedly much simpler) and thus I changed GetTable() within
the NovaGrid.js file to compensate for the differences. In addition, there
are buttons generated for the legend beneath the grid to delete whatever
by the group provided by the graph; in the case of configHoneyd nodes, 
delete all nodes by a certain profile. There is no change required to any 
other pi chart (to my knowledge), as if there's no deleteFunction given,
the button will not be rendered.