QUASAR: More work on scripts page
Table is generated with the Script Names, the profiles a script is linked to,
the path and parameters for the script, and a button to clear the script
from its associated profiles. 

Still need to:
Flesh out clear button functionality, not sure how to decouple the script
from the profiles; either have to remove the port altogether or just
have the port set to 'open' and remove the script, or whatever other 
behavior is desired. 
Add a way to add a script through a file upload dialog. This will be hard.