QUASAR: RemoveScriptFromProfiles successful
Right now, it just removes the port with the script attached from the profiles
using it. There are some additional things to be done for deletion:

Allow per-profile manipulation of the script, i.e. if a script is bound to one 
or more profiles, provide a way to select which profiles will have that port 
modified

Instead of deleting the port, instead just change the port to some 
predetermined default, and remove the scriptName and service components
(set to null string)