QUASAR: Fix bug with "do you want to restart honeyd?" prompts
There was a race condition where it wouldn't actually attempt to start
honeyd, because the stop honeyd command would return, but the honeyd
process was still shutting down. Instead of calling StartHaystack, which
checks first if it's arleady running, we just call Stop and Start, which can
result in a slight bit of overlap of the old and new honeyd process. The
overlap is mostly harmless though.