QUASAR: Remove Script
There is now a remove script  button within the /scripts page for each
row in the table that will perform the same steps as the 
ClearScriptFromProfiles method, and then remove the script from the 
m_scripts map within the HoneydConfiguration class. This will not remove
the script from the filesystem, but it does become non-selectable both on
the /scripts page as well as the profile page. 

Issue #325