QUASAR: Clear Script from Profiles now done on a per profile basis
If a script is linked to multiple profiles in the table, the lightbox will pop up
with the profiles next to checkboxes, with a checked checkbox indicating
that the accompanying profile is to have the script purged from its ports.

This option will not be presented for scripts with a single profile attached;
instead, it will immediately clear the script and reload the page.

Issue #325