MULT: More support for save/add/switching of configurations
Can now add a new configuration or clone an existing one. No conflicting
names are allowed. The default configuration is sovereign, and cannot be 
deleted. 

Can now clone a configuration off of a pre-existing configuration. 

Framework is now in place for allowing removal of configurations, still 
testing some workflow elements.

Tasks left: Provide a summary section s.t. on the page, when a new config
is selected, the user can get a decently abstracted overview of the 
composition of the configuration, and decide upon action from there.
There will be no way to edit individual components of the configurations
from this page; i.e. there may be a profile tree, but none of the elements will
be selectable. Merely meant to allow the user an 'educated guess' as to 
what's up with a given configuration and navigating quasar normally from 
there.

Issue #6