QUASAR: Remove many debug prints, fix remove configuration, add pi chart
Remove Configuration wasn't deleting the config entry from the 
configurations.txt file, meaning that if you made a new configuration
named 'asdf', removed it, and tried to either make a new configuration
with the same name, or clone an existing configuration and give it that same
name, the process wouldn't complete adequately. 

There's now a pi chart on the config summary screen that shows how many
nodes there are per profile within a configuration.

Issue #6