QUASAR: Saving profile wouldn't change UptimeMax value
There wasn't any logic within the now.SaveProfile method to allow for
the modification of the uptime max value, which ultimately meant that
even when the templates were saved, the uptime values would remain the
same (and thus, incorrect). This was causing some aesthetic problems on
the configuration summary page. Fixed now.

Issue #6 (tangentially)