QUASAR: Config summary page looks better
Reorganized the front page elements, and worked on the lightbox elements 
as well. Looks nice. Always going to be able to look better, but this is a good
point for it to be at.

Also fixed a timing issue for when the new configuration results were 
written to the haystack.config file, still need to test it out with honeyd,
but right now it doesn't pop the restart warning or anything.

Issue #6