MULT: Inital modifications to allow concurrent Quasar and Pulsar usage
Added a config variable for designating the port on the mothership.
Added Config methods to Set and Get the MASTER_UI_PORT variable value.
Within Quasar and Pulsar, the MASTER_UI_PORT variable is used to 
designate what port the websockets server is using to listen with 
on Pulsar. 

Note: Master_UI_Port's value designates the port to be used by Pulsar for
the user to interact with it through the browser. MASTER_UI_PORT + 1 is the
listening port. This has the possibility to create problems down the road
(i.e. someone sets Pulsar to 8079, and the listening server gets an
EADDRINUSE because 8080 is being used by Quasar for browser interaction)
and there is certainly work to be done, but this is the framework.