PULSAR: Add a small primer on Cron usage, and link to more help
One of the two styles of scheduled events is a cron string, used for 
precise recurrence events. However, there can be no expectation for 
users to have used or understand cron or its syntax prior to this. For now,
a lightbox will pop up the provides some simple information at first,
and a link to get more help (wikipedia for now, has a pretty robust 
explanation)