PULSAR: Change schedule elements to be simpler
Demanding either a cron string or a specific date/time format provides too 
much room for user error in the scheduling portion of Pulsar. Changing the
scheduling format element to just be the parameters required for the Date
element, which is what the node-scheduler takes as an arg for both single
and recurring events. Saving work to pull from other branches.