PULSAR: Fix the backend for scheduling, test
Currently, the user can choose either a one-off event to occur by giving
the day/month/year hour:minute desired for the event, or a recurring event
which takes a day of the week and hour:minute parameters. Tested both
(recurring on a smaller scale) and they seem to work. 

Work to be done:
Have to find some way to query the schedule object and determine when a 
one-off event has been completed; currently, the event will fire as intended,
but doesn't get removed from the UI element on the /listschedule page. 
Just have to do some research to determine how to get a list of the current
jobs from the schedule object and see if the lists are synced, removing any
extra within the server side array.