PULSAR: Add IE select .value workaround in Schedule page
As far as I can see, there were only two selects in use within Pulsar, this one
on the schedule page as well as one on the autoconfig page. Because IE 
doesn't adhere to DOM 1 in the instance of the select element (i.e. select
element having its own .value component to get the value of the currently
selected option element) one has to use the .options array and 
the currently selected index in order to procure the needed value.