PULSAR: Remove client from novaClients array when disconnected
Had it this way originally, decided not to go that route simply because 
Quasar instances could go up and down, and I preferred to allow the 
Pulsar user to see all of the clients that connected to pulsar, regardless
of whether they were connected or not. But given that our use case places
a medium sized limit on the amount of connected quasars, getting rid of 
them until they either reconnect or connect another a new name, I think 
this'll be fine.

Issue: #468