PULSAR: Roll back 108473153 and implement user-selected remove button
within the connectionsList element there's now a little red X that serves as 
a button to remove that connection from the list, the novaClients array,
and the selectable list at the top of most of the pages.

Wondering if using the button shouldn't blacklist that connection or 
something, it really doesn't do much against a connected Quasar instance
since it'll just reconnect within 5 seconds anyways.

Issue: #468