PULSAR: No more client disconnecting problems, only aesthetics
Removing the same client twice will no longer remove all clients; however,
the top border on the panels starts to clip off if all the clients are removed,
and the only way to get it back at this point that I can see is refreshing. 
I'll continue to work on it.