MULT: Add cancel button for autoconfig through QUASAR
There's now a cancel button in the lower left corner of the lightbox 
on the /autoConfig page. Currently it sends a SIGTERM to the autoconfig
process and redirects back the hhconfig page.

Things left to do:

Tell the user what happened, somehow. 
Make sure that the autoconfig tool works with the config management stuff


Issue: #493