MULT: Cancel button removes directory created by AddNewConfiguration
Autoconfig page now requires that a group name is given to create a new
haystack group when the autoconfig is run. Haystackautoconfig can be run
without one, placing a default groupname of 'autoconfig' + time(NULL) in
its place, but for the case of Quasar, it's going to be required.

Main functionality is in place, just testing for things like trying to create a 
group with a name already in use, etc.

Issue: #493