QUASAR: hhautoconfig was only showing eth1 when there was eth0 and eth1
hhautoconfig.jade was only showing two instances of eth1 when there was
an eth0 and an eth1. This was because of the timing of the callback 
function. It used variables outside its scope that had changed to incorrect
values by the time it needed to use them. Now passes through the required
values to GetSubnetFromInterface, all elements are created as they should
be, and the values shown within the UI are correct.

Close #502