QUASAR: Scoping problems with AddConfiguration inside of autoconfig
When it was getting called within quasar, the honeydconfiguration object
inside of Quasar was having access to the created group after it died
from the clicking of the cancel button. Now, the group gets created on the 
Quasar side, as well as inside of honeydConfiguration. There's no conflict,
as AddNewConfiguration checks for duplicates and handles them 
appropriately.

Close #493