AUTOCONFIG: Basic setup for adding autoconfig nodes by range
Accepts a range of the form ###.###.###.###-###.###.###.###, 
doesn't do much checking of the ips yet aside from making sure the range 
goes from low to high (i.e. start of range < end of range as given).

Creates the nodes with the correct addresses, note that the range is 
considered inclusive on both sides, meaning that if you have the range
1.1.1.1-1.1.1.4, you will get a node at 1.1.1.1, one at 1.1.1.2, one at 1.1.1.3, 
and one at 1.1.1.4.

#336