AUTOCONFIG: RemoveConfiguration calls need to be made within hconfig
The haystackautoconfig binary would create a group before, but if run
on the command line and execution was stopped (or there was an error)
the group would persist beyond its useful life. Now, there's a signal handler
for SIGINT and error handling RemoveConfigurations at areas where a 
set of configuration options would create an error condition.

#336