AUTOCONFIG: Accept comma-delimited ranges in Autoconfig tool
The -e flag now accepts the format
 ###.###.###.###-###.###.###.###,...

Seems pretty robust in handling malformed ips at this point, gave it some
error-producing ips and it either handles or shutsdown gracefully on each 
one.

I will say this: this feature (ip ranges in general) can be a major problem if 
the user is not careful in what they're doing. Say they mean to put in 
10.10.0.1-10.10.0.5 but put in 1.10.0.1-10.10.0.5. That range is going to take
a looooong time to fill.

close #527