QUASAR: Expose 'append-to' functionality within quasar
The append-to option is now accessed through a checkbox at the top of the 
autoconfig page. When it's checked, a new group is requested. This is the 
page default. When not checked, a select dropdown is shown with the 
groups capable of being appended to. This list excludes the 'default' group.

close #530