AUTOCONFIG: Autoconfig fails silently because LoadNmapXml fails silent
The return value for LoadNmapXml is a boolean regarding whether the 
function succeeded in its stated  purpose or not, but it was never checked,
creating the condition of nmap not being installed not creating a subnet.xml
file, LoadNmapXml realizing this and returning false, and autoconfig 
continuing as normal. Should be good now.

fixed #531