QUASAR: More WIP code for #536
Still getting random chinese characters in the portsetSelect select, and 
after redoGrid gets called for the test portset (which has been thoroughly
debugged at this point) changing back to the original portset fails to render
the grid. Curiously enough, if one goes into the dev console and highlights
the select, the page afterwards works as expected, displaying the correct
portsets, etc. Going to try this on my cubicle machine and see if the problem
persists, as well as try it with Chrome and Firefox, vs. Chromium as I'm using
now.