MULT: Allow renaming of portsets in Quasar
Right now, there's just a dumpy input and button combo beneath the 
edit port set select element on the editHoneydProfile page that, when used,
will traverse the ProfileTree and rename any portset matching oldName
to newName, then reload the page. 

One consideration is the need to traverse the tree: do I need to do this? Or 
can I just change the name for the individual profile and satisfy the ticket 
requirements?

#458