QUASAR: Profile tree div on honeydConfigManage forcing too many reflows
Trying to find a way to collapse the node-drawing process so that when the
list is actually finished it's appeneded to the DOM, as opposed to now,
when the new element gets appended in the first run of the recursive 
function, forcing a lot of reflows in the process and possibly bogging down
the element rewrite. WIP work to this end.