MULT: Hard shutdown option for Novad
Currently, is stop Novad returns false, pop HardShutdown. Looking at the
code for StopNovad, this seems reasonable, as it will return false only for 
a broken message line. May be overkill in some cases, but worse comes to 
worse, the user can just decline to use the prompt and click stop novad 
again.

#513