QUASAR: Select multiple nodes and change interface in configHoneydNodes
There is now a button that will take the selected nodes from the grid, 
prompt the user for their new interface as a string, and take the process
from there. If the string given does not match an interface on the current
list, warn the user, display the proper options, and try again. 

close #397