Fixed logic for pcap file over but could still be flawed
I didn't initially realize that there was a validation check for
the config file.  Since there is a validation check I simply
made the assumption that the person running novad with the pcap
file override knows what they're doing and will give a valid pcap
file.  I set the prefixes for both the READ_PCAP and PCAP_FILE to true