AUTOCONFIG: Fixed some cases where the lock file wasn't  being deleted
There were two cases where the lockfile wasn't being deleted, preventing
the proper use of the autoconfiguration tool after one use, which was failing
as well. Someone had changed the long-form name for the -i flag without
changing the parsing area, leading to a 100% exception rate for the 
Autoconfig, which wasn't failing the build because the generic error case 
returned 0 instead of a non-zero error code.