QUASAR: Allow right click on address text in NovaGrid
One can now right click on the address text in the NovaGrid and get the 
contextmenu. We have a pretty judicious use of divs hidden within the 
td elements that keep causing these problems, and will need to come up 
with a more general solution in order to compensate for arbitrary-depth
layers of divs within the tds that lead to proper resolution of the rendering
of the contextmenu, but for now it seems to work without problem.