QUASAR: After writing all settings for Config, reload config
Because the Config file was only ever getting written to in Quasar, and 
not read except for the initial call to Config::Inst() within Quasar's config
object, any saving of basic or advanced options would have no effect until 
Quasar was restarted. Now, at the end of configureNovaSave, the 
Config object reads from the file and updates its member variables to 
reflect the change.