QUASAR: Fix for #613
Clicking the inherit checkbox from not inherited to inherited was supposed
to default to the original profiles values, but it wasn't. Now it does. In 
addition, created a method to enforce min and max restrictions on number
type inputs, as the browser was not doing so as it has before. 

#613