QUASAR: Work on code style, fix a couple of bugs
Was using .attr('value', value) to set values and .attr('value') to get values
of attributes for a specific element; well, like with IE, selects and dropboxes 
don't want to play nicely, so we have to use a different method to produce
expected results. Luckily, it's just .val() and .val(value). Changed this on 
the honeydConfigManage page and in commonHoneydProfiles; I'm sure 
there's places I haven't hit yet that have a select and .attr(), but we'll call this
WIP for now and I'll continue along that path.