QUASAR: Tracking down Novad slow startup, fixing autoconfig bug
The GetNodeInterface method was returning null on a call to Autoconfig,
made sure that it's getting the correct value to pass before calling 
ShowAutoScan.

In addition, going through and trying to find problem with Novad starting up
really slowly; takes two clicks of start classifier in order to get UI responses
to Novad's status.